dependencies.config.ts file is renamed to init.ts
